requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Compress
python-dotenv
requests
Pillow
gunicorn
